Test creation (Add, Delete, Delete All)
